+++ CHANGELOG.md
@@ -9,17 +9,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 _Can be run using `npm run codetotal:beta`_
 
 - Global
+
   - Add animations to documentation
   - Add Google Tag Id for online doc
 
 - Front-end
+
   - Add linters logos
 
 - Back-End
+  - Async parsing of packages in SBOM
 
 ## [v0.4.0] - 2023-08-09
 
 - Global
+
   - Use latest version for production mode (by default), or beta version of MegaLinter to run CodeTotal in development mode
   - Documentation of docker-compose variables
   - docker-compose new variable: `REPOSITORY_TRUFFLEHOG_COMMAND_REMOVE_ARGUMENTS=--only-verified`
@@ -32,6 +36,7 @@ _Can be run using `npm run codetotal:beta`_
 ## [v0.3.0] - 2023-08-09
 
 - Global
+
   - Doc: display linter icons
 
 - Front-end
@@ -40,6 +45,7 @@ _Can be run using `npm run codetotal:beta`_
 ## [v0.2.0] - 2023-08-09
 
 - Global
+
   - Use `latest` codetotal-server docker image in production docker-compose.yml
   - Deploy online doc when any markdown file is updated
   - Factorize docker-compose
@@ -48,6 +54,7 @@ _Can be run using `npm run codetotal:beta`_
     - `npm run codetotal:dev` locally builds CodeTotal then runs it with beta version of MegaLinter
 
 - Front-end
+
   - Language selection in snippet analysis form
   - Language name added next to the icon in report header
   - "CodeTotal" text added to the footer